Add files via upload
Deep deterministic learning (DDL) is a systematic and pre-defined process to recognize the Cardiac diseases patterns of features after feature engineering. For operational investigations of DDL, the two datasets were used, first is a public dataset (MIT-BIH) that is attached in this repository and the second is an exclusive dataset of University of Malaya Medical center(UMMC) ( dataset of UMMC is not available in repository due to ethical agreement).
 For citation and close observations, here the follow the complete address of DDL published work 

Uzair Iqbal, Teh Ying Wah, Muhammad Habib Ur Rehman, Muhammad Imran, and Muhammad Shoaib, " Deep Deterministic Learning for Pattern Recognition of Different Cardiac Diseases through the Internet of Medical Things” in Journal of Medical Systems, 2018. DOI: 10.1007/s10916-018-1107-2